Fix: 게시판 삭제 확인 추가

+++ src/components/elements/BoardPage/Detail/_fragments/Main.tsx
@@ -141,11 +141,13 @@ function Main({ detail }: any) {
   }
 
   const deleteBoard = () => {
-    useGraphQL()
-      .deleteBoardDetail(detail.id)
-      .then(() => {
-        router.push('/board')
-      })
+    if (confirm('정말 삭제하시겠습니까?')) {
+      useGraphQL()
+        .deleteBoardDetail(detail.id)
+        .then(() => {
+          router.push('/board')
+        })
+    }
   }
 
   return (
